+++ src/lib/helpers.ts
@@ -1,6 +1,5 @@
 import type { User } from "@/types/api";
-
-import type { SelectInputOption } from "./types";
+import type { SelectInputOption } from "@/types/forms";
 
 /** Prefers the user's full name, falling back to their email. */
 export const getUserDisplayName = (user: User): string =>
